change correctly_pressed_letters and all_letter_pressed type from &mut i32 to &mut usize. Those should always be above 0

+++ src/event_handler.rs
@@ -17,8 +17,8 @@ pub fn on_keypress(
     c: char,
     did_mark_letter: &mut bool,
     cursor: &mut CursorPosition,
-    correctly_pressed_letters: &mut i32,
-    all_letter_pressed: &mut i32,
+    correctly_pressed_letters: &mut usize,
+    all_letter_pressed: &mut usize,
 ) -> bool {
     if *did_mark_letter {
         return true;
